4.21: added token authentication in blog DELETE

+++ controllers/blogs.js
@@ -28,8 +28,28 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!(request.token && decodedToken.id)) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(401).json({
+      error: `blog id ${request.params.id} not found`
+    })
+  }
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  } else {
+    return response.status(401).json({
+      error: 'delete request attempted by unauthorised user'
+    })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
